génération des cotations
génère des cotations mais ne fournit pas forcément les bonnes valeurs pour le moment.

+++ database/migrations/2019_02_04_152823_add_tables_cotations.php
@@ -16,9 +16,9 @@ class AddTablesCotations extends Migration
         Schema::create('cotations', function (Blueprint $table){
             $table->increments('id');
             $table->integer('crypto_id')->nullable();
-            $table->float('valeur')->nullable();
-            $table->double('cours')->nullable();
-            $table->float('evolution')->nullable();
+            $table->float('valeur',10,4)->nullable();
+            $table->float('cours',10,4)->nullable();
+            $table->float('evolution',10,4)->nullable();
             $table->datetime('date')->nullable();
             //$table->decimal('taux',5,2)->nullable();
             //$table->timestamps();
